tests: fix test_extern.nit for 64bits

+++ tests/test_extern_nit.c
@@ -1,3 +1,3 @@
-int foo(int s) { return 10; }
-int bar(int s, int i) { return i*2; }
-int baz(int s, int a) { return a; }
+long int foo(long int s) { return 10; }
+long int bar(long int s, long int i) { return i*2; }
+long int baz(long int s, long int a) { return a; }
